apps/backend/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.9.0
attrs==23.2.0
blinker==1.9.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.1.0
Flask-Cors==5.0.0
gitdb==4.0.11
GitPython==3.1.43
graphframes==0.6
graphx==0.0.4
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lark==1.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
networkx==3.4.2
nose==1.3.7
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
psutil==6.1.0
py4j==0.10.9.7
pyarrow==16.1.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-hcl2==7.2.0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.46.2
streamlit==1.36.0
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.34.2
Werkzeug==3.1.3
